Explicitly mention AppDaemon config file
I got confused (never used AppDaemon before) whether I have to add the config to Home Assistant's `configuration.yaml` or to the AppDaemon's config file. After a bit of frustration and a lot of reading, I found out this.

Maybe this happens to new users as well. That's why I thought its better to make this clearer